fix: apply gemini code reviews
- change Int type to Long for some big numbers
- change purchase logic to create ticket before withdraw
- fix critical bug in OutputView
- remove exitProcess(0) in handleMenu()

+++ lotto-project/src/main/kotlin/lotto/handler/DepositHandler.kt
@@ -12,7 +12,7 @@ class DepositHandler(
     override fun handle() {
         try {
             outputView.printGuidance("입금할 금액을 입력하세요: ")
-            val amount = inputView.inputPositive()
+            val amount = inputView.inputPositive().toLong()
             account.deposit(amount)
             outputView.printBalance(account.getAmount())
         } catch (e: IllegalArgumentException) {
